refactor(renderer): moved babel config into BabelLoader class
